Create Pencil Sketch with Python
we need to read the image in RGB format and then convert it into grayscale image. This will turn it into a classic black and white image. 
Then the next thing to do is invert the grayscale image also called negative image, this will be our inverted grayscale image. Inversion can be used to enhance details. 
Then we can finally create the pencil sketch by mixing the grayscale image with the inverted blurry image. This can be done by dividing the grayscale image by the inverted blurry image.